refactor: rename settings to wifiSettings
