+++ src/index.ts
@@ -1,8 +1,8 @@
 import * as android from './android'
 import * as automator from './automator'
 import * as device from './device'
-import * as settings from './settings'
+import * as wifiSettings from './wifi-settings'
 import logger from './logger'
 import poll from './poll'
 
-export { android, automator, device, settings, logger, poll }
+export { android, automator, device, wifiSettings, logger, poll }
